requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1 ; python_version >= '3.8'
django==5.2.1 ; python_version >= '3.10'
django-filter==25.1 ; python_version >= '3.9'
djangorestframework==3.16.0 ; python_version >= '3.9'
djangorestframework-simplejwt==5.5.0 ; python_version >= '3.9'
drf-nested-routers==0.94.1 ; python_version >= '3.8'
faker==37.1.0 ; python_version >= '3.9'
pyjwt==2.9.0 ; python_version >= '3.8'
sqlparse==0.5.3 ; python_version >= '3.8'
tzdata==2025.2 ; sys_platform == 'win32'
